Move sincedb logic to DDB
  In short, sincedb doesn't work in S3 (where it was copied from), and
works even worse here.

  The original model stores a timestamp, and uses that as the watermark -
items older than it are ignored, newer than it are merged, but this has
a number of problems.

  First, we're not guaranteed all logging is _sequentially_ complete -
it's quite possible to read page 1,2,4 and miss the content of page 3
until reading later - simply because it's not there.

  Second, a single timestamp doesn't fairly represent our position across
multiple loggroups and logstreams.

  Third, the state is preserved to the local filesystem, which works, but
in my environment, the container is created/destroyed as resources are
needed/available, and the local filesystem isn't preserved.

  Forth, re: state, there's no way for multiple workers to interlace using
the local filesystem (without filesharing semantics).

  Fifth, the recursive function calls are leaking memory like a scieve,
and that led to crashes, leading to the discovery and workarounds in (1-4).

We now store state in a table, and extend the featureset to optionally support
expunging old datasets.